+++ map_gen/Diggy/Feature/MarketExchange.lua
@@ -11,10 +11,11 @@ local Debug = require 'map_gen.Diggy.Debug'
 local Template = require 'map_gen.Diggy.Template'
 local Global = require 'utils.global'
 local Game = require 'utils.game'
-local MarketUnlockables = require 'map_gen.Diggy.MarketUnlockables'
-local calculate_level = MarketUnlockables.calculate_level
 local insert = table.insert
+local force_control = require 'features.force_control'
+local Experience = require 'map_gen.Diggy.Feature.Experience'
 local max = math.max
+local floor = math.floor
 local utils = require 'utils.utils'
 local prefix = '## - '
 
@@ -23,114 +24,20 @@ local MarketExchange = {}
 
 local config = {}
 
-local stone_tracker = {
-    stone_sent_to_surface = 0,
-    previous_stone_sent_to_surface = 0,
-    current_level = 0,
-}
-
-local stone_collecting = {
-    initial_value = 0,
-    active_modifier = 0,
-    research_modifier = 0,
-    market_modifier = 0,
-}
-
-local mining_efficiency = {
-    active_modifier = 0,
-    research_modifier = 0,
-    market_modifier = 0,
-}
-
-local inventory_slots = {
-    active_modifier = 0,
-    research_modifier = 0,
-    market_modifier = 0,
-}
-
-Global.register({
-    stone_collecting = stone_collecting,
-    stone_tracker = stone_tracker,
-    mining_efficiency = mining_efficiency,
-    inventory_slots = inventory_slots,
-}, function(tbl)
-    stone_collecting = tbl.stone_collecting
-    stone_tracker = tbl.stone_tracker
-    mining_efficiency = tbl.mining_efficiency
-    inventory_slots = tbl.inventory_slots
-end)
-
-local function send_stone_to_surface(total)
-    stone_tracker.previous_stone_sent_to_surface = stone_tracker.stone_sent_to_surface
-    stone_tracker.stone_sent_to_surface = stone_tracker.stone_sent_to_surface + total
-end
-
 local on_market_timeout_finished = Token.register(function(params)
     Template.market(params.surface, params.position, params.player_force, {})
-
-    local tiles = {}
-    for _, position in pairs(params.void_chest_tiles) do
-        insert(tiles, {name = 'tutorial-grid', position = position})
-    end
-
-    params.surface.set_tiles(tiles)
 end)
 
-local function update_mining_speed(force)
-    -- remove the current buff
-    local old_modifier = force.manual_mining_speed_modifier - mining_efficiency.active_modifier
-
-    -- update the active modifier
-    mining_efficiency.active_modifier = mining_efficiency.research_modifier + mining_efficiency.market_modifier
-
-    -- add the new active modifier to the non-buffed modifier
-    force.manual_mining_speed_modifier = old_modifier + mining_efficiency.active_modifier
-end
-
-local function update_inventory_slots(force)
-    -- remove the current buff
-    local old_modifier = force.character_inventory_slots_bonus - inventory_slots.active_modifier
-
-    -- update the active modifier
-    inventory_slots.active_modifier = inventory_slots.research_modifier + inventory_slots.market_modifier
-
-    -- add the new active modifier to the non-buffed modifier
-    force.character_inventory_slots_bonus = old_modifier + inventory_slots.active_modifier
-end
-
-local function update_stone_collecting()
-    -- remove the current buff
-    local old_modifier = stone_collecting.initial_value - stone_collecting.active_modifier
-
-    -- update the active modifier
-    stone_collecting.active_modifier = stone_collecting.research_modifier + stone_collecting.market_modifier
-
-    -- add the new active modifier to the non-buffed modifier
-    stone_collecting.initial_value = old_modifier + stone_collecting.active_modifier
-end
-
-
---Handles the updating of market items when unlocked, also handles the buffs
-local function update_market_contents(market)
-    if (stone_tracker.previous_stone_sent_to_surface == stone_tracker.stone_sent_to_surface) then
-        return
-    end
-
-    local should_update_mining_speed = false
-    local should_update_inventory_slots = false
-    local should_update_stone_collecting = false
+---Updates market content with new items if they are to be unlocked
+---Primarily used by the force control system at every level up
+---@param market LuaEntity The market to be updated
+---@param force LuaForce the force which the unlocking requirement should be based of
+function MarketExchange.update_market_contents(market, force)
     local add_market_item
-    local old_level = stone_tracker.current_level
-    local print = game.print
     local item_unlocked = false
 
-    if (calculate_level(stone_tracker.current_level+1) <= stone_tracker.stone_sent_to_surface) then
-        stone_tracker.current_level = stone_tracker.current_level + 1
-    end
-
     for _, unlockable in pairs(config.unlockables) do
-        local stone_unlock = calculate_level(unlockable.level)
-        local is_in_range = stone_unlock > stone_tracker.previous_stone_sent_to_surface and stone_unlock <= stone_tracker.stone_sent_to_surface
+        local is_in_range = force_control.get_force_data(force).current_level == unlockable.level
 
         -- only add the item to the market if it's between the old and new stone range
         if (is_in_range and unlockable.type == 'market') then
@@ -153,84 +60,18 @@ local function update_market_contents(market)
 
         end
     end
-
-    MarketExchange.update_gui()
-
-    if (old_level < stone_tracker.current_level) then
-        if item_unlocked then
-            print(prefix..'We have reached level ' .. stone_tracker.current_level .. '! New items are available from the market!')
-        else
-            print(prefix..'We have reached level ' .. stone_tracker.current_level .. '!')
-        end
-        for _, buffs in pairs(config.buffs) do
-            if (buffs.prototype.name == 'mining_speed') then
-                local value = buffs.prototype.value
-                Debug.print('Mining Foreman: Increased mining speed by ' .. value .. '%!')
-                should_update_mining_speed = true
-                mining_efficiency.market_modifier = mining_efficiency.market_modifier + (value / 100)
-            elseif (buffs.prototype.name == 'inventory_slot') then
-                local value = buffs.prototype.value
-                Debug.print('Mining Foreman: Increased inventory slots by ' .. value .. '!')
-                should_update_inventory_slots = true
-                inventory_slots.market_modifier = inventory_slots.market_modifier + value
-            elseif (buffs.prototype.name == 'stone_automation') then
-                local value = buffs.prototype.value
-                if (stone_tracker.current_level == 1) then
-                    print('Mining Foreman: We can now automatically send stone to the surface from a chest below the market!')
-                else
-                    Debug.print('Mining Foreman: We can now automatically send ' .. value .. ' more stones!')
-                end
-                should_update_stone_collecting = true
-                stone_collecting.market_modifier = stone_collecting.market_modifier + value
-            end
-        end
-    end
-
-    local force
-
-    if (should_update_mining_speed) then
-        force = force or game.forces.player
-        update_mining_speed(force)
-    end
-
-    if (should_update_inventory_slots) then
-        force = force or game.forces.player
-        update_inventory_slots(force)
-    end
-
-    if (should_update_stone_collecting) then
-        update_stone_collecting()
-    end
-end
-
-local function on_research_finished(event)
-    local force = game.forces.player
-    local current_modifier = mining_efficiency.research_modifier
-    local new_modifier = force.mining_drill_productivity_bonus * config.mining_speed_productivity_multiplier * 0.5
-
-    if (current_modifier == new_modifier) then
-        -- something else was researched
-        return
-    end
-
-    mining_efficiency.research_modifier = new_modifier
-    inventory_slots.research_modifier = force.mining_drill_productivity_bonus * 50 -- 1 per level
-    stone_collecting.research_modifier = force.mining_drill_productivity_bonus * 1250 -- 25 per level
-
-    update_inventory_slots(force)
-    update_mining_speed(force)
-    update_stone_collecting()
 end
 
 local function redraw_title(data)
-    data.frame.caption = utils.comma_value(stone_tracker.stone_sent_to_surface) .. ' ' .. config.currency_item .. ' sent to the surface'
+    local force_data = force_control.get_force_data('player')
+    data.frame.caption = utils.comma_value(force_data.total_experience) .. ' total experience earned!'
 end
 
 local function get_data(unlocks, stone, type)
     local result = {}
 
     for _, data in pairs(unlocks) do
-        if calculate_level(data.level) == stone and data.type == type then
+        if data.level == stone and data.type == type then
             insert(result, data)
         end
     end
@@ -259,23 +100,14 @@ local function redraw_heading(data, header)
 end
 
 local function redraw_progressbar(data)
-
+    local force_data = force_control.get_force_data('player')
     local flow = data.market_progressbars
     Gui.clear(flow)
 
-    -- progress bar for next level
-    local act_stone = (stone_tracker.current_level ~= 0) and calculate_level(stone_tracker.current_level) or 0
-    local next_stone = calculate_level(stone_tracker.current_level+1)
-
-    local range = next_stone - act_stone
-    local sent = stone_tracker.stone_sent_to_surface - act_stone
-    local percentage = (math.floor((sent / range)*1000))*0.001
-    percentage = (percentage < 0) and (percentage*-1) or percentage
-
-    apply_heading_style(flow.add({type = 'label', tooltip = 'Currently at level: ' .. stone_tracker.current_level .. '\nNext level at: ' .. utils.comma_value(next_stone) ..'\nRemaining stone: ' .. utils.comma_value(range - sent), name = 'Diggy.MarketExchange.Frame.Progress.Level', caption = 'Progress to next level:'}).style)
-    local level_progressbar = flow.add({type = 'progressbar', tooltip = percentage * 100 .. '% stone to next level'})
+    apply_heading_style(flow.add({type = 'label', tooltip = 'Currently at level: ' .. force_data.current_level .. '\nNext level at: ' .. utils.comma_value((force_data.total_experience - force_data.current_experience) + force_data.experience_level_up_cap) ..' xp\nRemaining xp: ' .. utils.comma_value(force_data.experience_level_up_cap - force_data.current_experience), name = 'Diggy.MarketExchange.Frame.Progress.Level', caption = 'Progress to next level:'}).style)
+    local level_progressbar = flow.add({type = 'progressbar', tooltip = floor(force_data.experience_percentage*100)*0.01 .. '% xp to next level'})
     level_progressbar.style.width = 350
-    level_progressbar.value = percentage
+    level_progressbar.value = force_data.experience_percentage * 0.01
 end
 
 local function redraw_table(data)
@@ -296,10 +128,10 @@ local function redraw_table(data)
 
     -- create table
     for i = 1, #config.unlockables do
-        if calculate_level(config.unlockables[i].level) ~= last_stone then
+        if config.unlockables[i].level ~= last_stone then
 
             -- get items and buffs for each stone value
-            items = get_data(config.unlockables, calculate_level(config.unlockables[i].level), 'market')
+            items = get_data(config.unlockables, config.unlockables[i].level, 'market')
 
             -- get number of rows
             number_of_rows = max(#buffs, #items)
@@ -311,8 +143,7 @@ local function redraw_table(data)
                 local level = item.level
 
                 -- 1st column
-                result[6] = calculate_level(level)
-                result[1] = 'Level ' ..level
+                result[1] = level
                 -- 3rd column
                 if items[j] ~= nil then
                     result[3] = '+ ' .. item.prototype.name
@@ -337,23 +168,23 @@ local function redraw_table(data)
         end
 
         -- save lastStone
-        last_stone = calculate_level(config.unlockables[i].level)
+        last_stone = config.unlockables[i].level
     end
 
     -- print table
     for _, unlockable in pairs(row) do
-        local is_unlocked = unlockable[6] <= stone_tracker.stone_sent_to_surface
+        local is_unlocked = unlockable[1] <= force_control.get_force_data('player').current_level
         local list = market_scroll_pane.add {type = 'table', column_count = 2 }
 
         list.style.horizontal_spacing = 16
 
         local caption = ''
         if unlockable[4] ~= true then
-            caption = unlockable[1]
+            caption = 'Level ' .. unlockable[1]
         else
             caption = ''
         end
-        local tag_stone = list.add {type = 'label', name = tag_label_stone, caption = caption}
+        local tag_stone = list.add {type = 'label', name = tag_label_stone, caption = caption, tooltip = 'XP: ' .. utils.comma_value(Experience.calculate_level_xp(unlockable[1]))}
         tag_stone.style.minimal_width = 100
 
         local tag_items = list.add {type = 'label', name = tag_label_item, caption = unlockable[3]}
@@ -378,30 +209,27 @@ local function redraw_buff(data) --! Almost equals to the redraw_table() functio
     local buff_scroll_pane = data.buff_scroll_pane
     Gui.clear(buff_scroll_pane)
 
-    local buffs = {}
-    local number_of_rows = 0
+    local buffs = Experience.get_buffs()
     local row = {}
-
-    for i = 1, #config.buffs do
-        -- get items and buffs for each stone value
-        buffs = config.buffs
-
+    local i = 0
+    for k, v in pairs(buffs) do
+        i = i + 1
         local result = {}
 
         -- 1st column
         result[1] = 'All levels'
 
         -- 2nd column
-        if buffs[i].prototype.name == 'mining_speed' then
-            result[2] = '+ '.. buffs[i].prototype.value .. '% mining speed'
-        elseif buffs[i].prototype.name == 'inventory_slot' then
-            if buffs[i].prototype.value > 1 then
-                result[2] = '+ '.. buffs[i].prototype.value .. ' inventory slots'
+        if k == 'mining_speed' then
+            result[2] = '+ '.. v.value .. '% mining speed'
+        elseif k == 'inventory_slot' then
+            if v.value > 1 then
+                result[2] = '+ '.. v.value .. ' inventory slots'
             else
-                result[2] = '+ '.. buffs[i].prototype.value .. ' inventory slot'
+                result[2] = '+ '.. v.value .. ' inventory slot'
             end
-        elseif buffs[i].prototype.name == 'stone_automation' then
-            result[2] = '+ '.. buffs[i].prototype.value .. ' stones automatically sent'
+        elseif k == 'health_bonus' then
+            result[2] = '+ '.. v.value .. ' max health'
         else
             result[2] = 'Description missing: unknown buff. Please contact admin'
         end
@@ -437,34 +265,29 @@ local function redraw_buff(data) --! Almost equals to the redraw_table() functio
     end
 end
 
-local function on_market_item_purchased(event)
-    if (1 ~= event.offer_index) then
+---Interface for force control system to access the primary market
+---@return LuaEntity the primary market (The one at spawn)
+function MarketExchange.get_market()
+
+    local markets = game.surfaces.nauvis.find_entities_filtered({name = 'market', position = config.market_spawn_position, limit = 1})
+
+    if (#markets == 0) then
+        Debug.print_position(config.market_spawn_position, 'Unable to find a market')
         return
     end
 
-    local sum = config.stone_to_surface_amount * event.count
-    Game.print_player_floating_text(event.player_index, '-' .. sum .. ' stone', {r = 0.6, g = 0.55, b = 0.42})
-
-    send_stone_to_surface(sum)
-    update_market_contents(event.market)
+    return markets[1]
 end
 
 local function on_placed_entity(event)
     local market = event.entity
-    if ('market' ~= market.name) then
+    if 'market' ~= market.name then
         return
     end
-
-    market.add_market_item({
-        price = {{config.currency_item, 50}},
-        offer = {type = 'nothing', effect_description = 'Send ' .. config.stone_to_surface_amount .. ' ' .. config.currency_item .. ' to the surface. To see the overall progress and rewards, click the market button in the menu.'}
-    })
-
-    update_market_contents(market)
 end
 
 function MarketExchange.get_extra_map_info(config)
-    return 'Market Exchange, trade your stone or send it to the surface'
+    return 'Market Exchange, come make a deal at the foreman\'s shop'
 end
 
 local function toggle(event)
@@ -530,6 +353,7 @@ Gui.on_custom_close('Diggy.MarketExchange.Frame', function (event)
     event.element.destroy()
 end)
 
+---Updates the market progress gui for every player that has it open
 function MarketExchange.update_gui()
     for _, p in ipairs(game.connected_players) do
         local frame = p.gui.left['Diggy.MarketExchange.Frame']
@@ -537,7 +361,6 @@ function MarketExchange.update_gui()
         if frame and frame.valid then
             local data = {player = p, trigger = 'update_gui'}
             toggle(data)
-            --toggle(data)
         end
     end
 end
@@ -547,10 +370,9 @@ function MarketExchange.on_init()
         surface = game.surfaces.nauvis,
         position = config.market_spawn_position,
         player_force = game.forces.player,
-        void_chest_tiles = config.void_chest_tiles,
     })
 
-    update_mining_speed(game.forces.player)
+
 end
 
 --[[--
@@ -559,91 +381,10 @@ end
 function MarketExchange.register(cfg)
     config = cfg
 
-    Event.add(defines.events.on_research_finished, on_research_finished)
-    Event.add(defines.events.on_market_item_purchased, on_market_item_purchased)
+    --Events
     Event.add(Template.events.on_placed_entity, on_placed_entity)
     Event.add(defines.events.on_player_created, on_player_created)
-
-    local x_min
-    local y_min
-    local x_max
-    local y_max
-
-    for _, position in pairs(config.void_chest_tiles) do
-        local x = position.x
-        local y = position.y
-
-        if (nil == x_min or x < x_min) then
-            x_min = x
-        end
-
-        if (nil == x_max or x > x_max) then
-            x_max = x
-        end
-
-        if (nil == y_min or y < y_min) then
-            y_min = y
-        end
-
-        if (nil == y_max or y > y_max) then
-            y_max = y
-        end
-    end
-
-    local area = {{x_min, y_min}, {x_max + 1, y_max + 1}}
-    local message_x = (x_max + x_min) * 0.5
-    local message_y = (y_max + y_min) * 0.5
-
-    Event.on_nth_tick(config.void_chest_frequency, function ()
-        local send_to_surface = 0
-        local surface = game.surfaces.nauvis
-        local find_entities_filtered = surface.find_entities_filtered
-        local chests = find_entities_filtered({area = area, type = {'container', 'logistic-container'}})
-        local to_fetch = stone_collecting.active_modifier
-
-        for _, chest in pairs(chests) do
-            local chest_contents = chest.get_inventory(defines.inventory.chest)
-            local stone_in_chest = chest_contents.get_item_count(config.currency_item)
-            local delta = to_fetch
-
-            if (stone_in_chest < delta) then
-                delta = stone_in_chest
-            end
-
-            if (delta > 0) then
-                chest_contents.remove({name = config.currency_item, count = delta})
-                send_to_surface = send_to_surface + delta
-            end
-        end
-
-        if (send_to_surface == 0) then
-            if (0 == to_fetch) then
-                return
-            end
-
-            local message = 'Missing chests below market'
-            if (#chests > 0) then
-                message = 'No stone in chests found'
-            end
-
-            Game.print_floating_text(surface, {x = message_x, y = message_y}, message, { r = 220, g = 100, b = 50})
-            return
-        end
-
-        local markets = find_entities_filtered({name = 'market', position = config.market_spawn_position, limit = 1})
-
-        if (#markets == 0) then
-            Debug.print_position(config.market_spawn_position, 'Unable to find a market')
-            return
-        end
-
-        local message = send_to_surface .. ' stone sent to the surface'
-
-        Game.print_floating_text(surface, {x = message_x, y = message_y}, message, { r = 0.6, g = 0.55, b = 0.42})
-
-        send_stone_to_surface(send_to_surface)
-        update_market_contents(markets[1])
-    end)
+    Event.on_nth_tick(61, MarketExchange.update_gui)
 end
 
 return MarketExchange